Added changes for Debian (+ PHP version check)
Debian does not have the command "add-apt-repository" pre-installed, so executing "sudo apt-get install software-properties-common -y" is a requirement for the script to work on all debian distributions besides Ubuntu. With this little addition the script won't cancel at this point and will work for Debian 8 and Ubuntu 14.04 (tested).

Also added a check for the PHP version, so the repository "ppa:ondrej/php5-5.6" will not be added to the list, if a newer version of PHP is already installed on the system. If the repository is not needed, then the requirements for the "add-apt-repository" command won't be added unnecessarily either. The value of the PHPVER variable will be "false" both if PHP is not installed or a lower version than "5.6" is detected and then install the "add-apt-repository" command and add the repository to the list.